Day 6

+++ 6_test.go
@@ -3,7 +3,7 @@ package main
 import "testing"
 
 func TestDay6Exec80(t *testing.T) {
-	expected := 5934
+	expected := int64(5934)
 	actual := Day6Exec("inputs/6_sample.txt", 80)
 
 	if expected != actual {
@@ -12,7 +12,7 @@ func TestDay6Exec80(t *testing.T) {
 }
 
 func TestDay6Exec256(t *testing.T) {
-	expected := 26984457539
+	expected := int64(26984457539)
 	actual := Day6Exec("inputs/6_sample.txt", 256)
 
 	if expected != actual {
